Add search page with header and a few aesthetic improvements
Implemented a search page, though it currently cannot be navigated to from the home page, with an acceptable header similar to Google's own on their search page.

Also implemented a few aesthetic improvements overall, most notably the box-shadow on the search box. This includes some CSS rules that weren't there before so that they don't break the home page while I work on the search page.