Add search results, footer and access to search from index
Added example search results with made-up data. A footer was also added and both of these additions should be styled correctly, at least on Firefox, with any luck.

Also enabled search.html to be accessible from index.html - simply press enter in the search field and it will take the user to the search page.